requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
Django==5.2.6
django-allauth==65.11.2
django-cors-headers==4.8.0
django-ipware==7.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
python-ipware==3.0.0
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
